requirements.txt
asgiref
Django
django-cors-headers
django-dotenv
django-filter
djangorestframework
djangorestframework-simplejwt
gunicorn
psycopg2
PyJWT
python-dotenv
pytz
sqlparse
tzdata
